backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
openai
python-multipart
websockets==12.0
Authlib
cryptography
itsdangerous
httpx
python-dotenv
sqlalchemy
aiosqlite
mysql-connector-python
